pipeline/requirements.txt
feedparser
psycopg2-binary
langchain
langchain-openai
langchain-google-genai
langchain-groq
python-dotenv
pydantic-settings
tenacity
asyncio
python-dateutil
beautifulsoup4
sentence-transformers
